requirements.txt
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==2.0.0
cryptography==41.0.6
Deprecated==1.2.14
idna==3.4
pycparser==2.21
PyGithub==1.56
PyJWT==2.4.0
PyNaCl==1.5.0
requests==2.31.0
urllib3==1.26.18
wrapt==1.15.0